+++ src/cli/plugin/index.ts
@@ -1,7 +1,9 @@
 import type { CommandModule } from "yargs";
 import type yargs from "yargs";
+import { keygenCommand } from "./keygen";
 import { packCommand } from "./pack";
 import { infoCommand } from "./info";
+import { uploadCommand } from "./upload";
 import { setStability } from "../stability";
 import { initCommand } from "./init";
 
@@ -12,7 +14,9 @@ const describe = "Commands for kintone plugin";
 const builder = (args: yargs.Argv) =>
   args
     .command(infoCommand)
+    .command(keygenCommand)
     .command(packCommand)
+    .command(uploadCommand)
     .command(initCommand)
     .demandCommand();